requirements.txt
faust-streaming==0.9.0